+++ h3ron/src/util.rs
@@ -1,11 +1,12 @@
 use geo_types::{Coordinate, LineString, Point};
 
-use h3ron_h3_sys::{degsToRads, GeoCoord};
+use h3ron_h3_sys::GeoCoord;
 
-pub(crate) unsafe fn coordinate_to_geocoord(c: &Coordinate<f64>) -> GeoCoord {
+#[inline(always)]
+pub(crate) fn coordinate_to_geocoord(c: &Coordinate<f64>) -> GeoCoord {
     GeoCoord {
-        lat: degsToRads(c.y),
-        lon: degsToRads(c.x),
+        lat: c.y.to_radians(),
+        lon: c.x.to_radians(),
     }
 }
 
@@ -13,9 +14,10 @@ pub(crate) unsafe fn linestring_to_geocoords(ls: &LineString<f64>) -> Vec<GeoCoo
     ls.points_iter().map(|p| point_to_geocoord(&p)).collect()
 }
 
-pub(crate) unsafe fn point_to_geocoord(pt: &Point<f64>) -> GeoCoord {
+#[inline(always)]
+pub(crate) fn point_to_geocoord(pt: &Point<f64>) -> GeoCoord {
     GeoCoord {
-        lat: degsToRads(pt.y()),
-        lon: degsToRads(pt.x()),
+        lat: pt.y().to_radians(),
+        lon: pt.x().to_radians(),
     }
 }